dev-requirements.txt
pytest-ckan
pytest-factoryboy
responses
aioresponses
pytest-asyncio
